+++ ReleaseNotes.md
@@ -1,3 +1,8 @@
+## Changelog DistributionGroupMigration v0.1.1
+
+1. Minor bug fixes for ```Initialize-OnPremDSTGroupToCloud``` and add validation for Accepted Domains in Exchange Online.
+
+
 ## Changelog DistributionGroupMigration v0.1.0
 
 1. Initial release of the DistributionGroupMigration module.